requirements_with_apis.txt
# AI Content Factory - API Dependencies
flask==3.1.2
requests==2.31.0
python-dotenv==1.0.0
sqlite3

# Data Sources (ฟรี)
pytrends==4.9.2
google-api-python-client==2.108.0

# AI Services
groq==0.4.1
openai==1.3.7

# Optional (สำหรับ advanced features)
beautifulsoup4==4.12.2
pandas==2.1.4
matplotlib==3.8.2
